+++ CPACqc/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.2.12'
+__version__ = '0.2.13'
 __author__ = 'Biraj Shrestha'
 __email__ = 'birajstha@gmail.com'
 __description__ = 'A package to view Nifti files in a BIDS dataset and generate QC plots.'
