--- CPACqc/config.py
@@ -1,2 +1,0 @@
-
-# TODO Add default configs here
