--- CPACqc/logging/log.py
@@ -1,25 +1,0 @@
-import logging
-import os
-
-logger = None
-
-def setup_logger():
-    # setup logger
-    logger = logging.getLogger(__name__)
-    logger.setLevel(logging.INFO)
-
-    # Create a file handler
-    log_file = os.path.join(os.getcwd(), 'CPACqc.log')
-    file_handler = logging.FileHandler(log_file)
-    file_handler.setLevel(logging.INFO)
-
-    # Create a logging format
-    formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-    file_handler.setFormatter(formatter)
-
-    # Add the handlers to the logger
-    logger.addHandler(file_handler)
-    
-    return logger
-
-logger = setup_logger()
