+++ CPACqc/utils/logging/log.py
@@ -1,0 +1,53 @@
+from CPACqc.services.logging_service import LoggerService
+import logging
+
+class ConsoleLogger(LoggerService):
+    """
+    A simple console logger that implements the LoggerAdapter interface.
+    """
+    def __init__(self, name: str):
+        self.logger = logging.getLogger(name)
+        self.logger.setLevel(logging.DEBUG)
+        ch = logging.StreamHandler()
+        ch.setLevel(logging.DEBUG)
+        formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+        ch.setFormatter(formatter)
+        self.logger.addHandler(ch)
+
+    def info(self, message: str):
+        self.logger.info(message)
+
+    def error(self, message: str):
+        self.logger.error(message)
+
+    def debug(self, message: str):
+        self.logger.debug(message)
+
+    def warning(self, message: str):
+        self.logger.warning(message)
+
+class FileLogger(LoggerService):
+    """
+    A file logger that implements the LoggerAdapter interface.
+    """
+    def __init__(self, name: str, filename: str):
+        self.logger = logging.getLogger(name)
+        self.logger.setLevel(logging.DEBUG)
+        fh = logging.FileHandler(filename)
+        fh.setLevel(logging.DEBUG)
+        formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+        fh.setFormatter(formatter)
+        self.logger.addHandler(fh)
+
+    def info(self, message: str):
+        self.logger.info(message)
+
+    def error(self, message: str):
+        self.logger.error(message)
+
+    def debug(self, message: str):
+        self.logger.debug(message)
+
+    def warning(self, message: str):
+        self.logger.warning(message)
+
