+++ CPACqc/utils/tabler/custom_bids2table/entities.py
