+++ CPACqc/utils/templates/script.js
