+++ pdm.lock
@@ -1,0 +1,11 @@
+# This file is @generated by PDM.
+# It is not intended for manual editing.
+
+[metadata]
+groups = ["default"]
+strategy = ["inherit_metadata"]
+lock_version = "4.5.0"
+content_hash = "sha256:59768564dce723c1376c0155e78c43bf1446ec8a8b1b91d711920aa9ec0a4faf"
+
+[[metadata.targets]]
+requires_python = ">=3.10"
